Split autodiscovery methods into gradle builds
This commit should be faily self-explainatory. It splits the
three autodiscovery methods into their own gradle builds so
they can be plugged in and out easily. This commit doesn't
necessarily make the latter possible though.

Dependency graph should look something like this now:

  app:ui
  \
   \app:autodiscovery:api
    |
    |\app:autodiscovery:providersxml
    |
    |\app:autodiscovery:srvrecords
    |
     \app:autodiscovery:thunderbird
