+++ app/autodiscovery/api/build.gradle
