+++ app/autodiscovery/providersxml/src/main/res/xml/providers.xml
