+++ app/autodiscovery/providersxml/src/test/resources/mockito-extensions/org.mockito.plugins.MockMaker
