--- app/autodiscovery/src/main/java/com/fsck/k9/autodiscovery/ConnectionSettings.kt
@@ -1,5 +1,0 @@
-package com.fsck.k9.autodiscovery
-
-import com.fsck.k9.mail.ServerSettings
-
-data class ConnectionSettings(val incoming: ServerSettings, val outgoing: ServerSettings)
