+++ app/autodiscovery/srvrecords/src/main/java/com/fsck/k9/autodiscovery/srvrecords/MiniDnsSrvResolver.kt
