+++ app/ui/src/main/java/com/fsck/k9/activity/setup/AccountSetupBasics.java
@@ -23,10 +23,9 @@ import com.fsck.k9.Preferences;
 import com.fsck.k9.account.AccountCreator;
 import com.fsck.k9.activity.K9Activity;
 import com.fsck.k9.activity.setup.AccountSetupCheckSettings.CheckDirection;
-import com.fsck.k9.autodiscovery.ConnectionSettings;
-import com.fsck.k9.autodiscovery.DiscoveredServerSettings;
-import com.fsck.k9.autodiscovery.DiscoveryResults;
-import com.fsck.k9.autodiscovery.DiscoveryTarget;
+import com.fsck.k9.autodiscovery.api.DiscoveredServerSettings;
+import com.fsck.k9.autodiscovery.api.DiscoveryResults;
+import com.fsck.k9.autodiscovery.api.DiscoveryTarget;
 import com.fsck.k9.autodiscovery.providersxml.ProvidersXmlDiscovery;
 import com.fsck.k9.backend.BackendManager;
 import com.fsck.k9.helper.EmailHelper;
@@ -36,6 +35,7 @@ import com.fsck.k9.mail.ConnectionSecurity;
 import com.fsck.k9.mail.ServerSettings;
 import com.fsck.k9.preferences.Protocols;
 import com.fsck.k9.ui.R;
+import com.fsck.k9.ui.ConnectionSettings;
 import com.fsck.k9.view.ClientCertificateSpinner;
 import com.fsck.k9.view.ClientCertificateSpinner.OnClientCertificateChangedListener;
 import timber.log.Timber;
