+++ app/ui/src/main/java/com/fsck/k9/ui/ConnectionSettings.kt
@@ -1,0 +1,5 @@
+package com.fsck.k9.ui
+
+import com.fsck.k9.mail.ServerSettings
+
+data class ConnectionSettings(val incoming: ServerSettings, val outgoing: ServerSettings)
