+++ settings.gradle
@@ -5,7 +5,10 @@ include ':app:core'
 include ':app:storage'
 include ':app:crypto-openpgp'
 include ':app:testing'
-include ':app:autodiscovery'
+include ':app:autodiscovery:api'
+include ':app:autodiscovery:providersxml'
+include ':app:autodiscovery:srvrecords'
+include ':app:autodiscovery:thunderbird'
 include ':mail:common'
 include ':mail:testing'
 include ':mail:protocols:imap'
